+++ docs/changelog.md
@@ -2,6 +2,16 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format.
 
+## [1.7.0] -- 2024-01-26
+
+### Added
+- `--portable` flag to `looper report` to create a portable version of the html report
+- `--lump-j` allows grouping samples into a defined number of jobs
+
+### Changed
+- `--lumpn` is now `--lump-n`
+- `--lump` is now `--lump-s`
+
 ## [1.6.0] -- 2023-12-22
 
 ### Added